+++ src/lib/settings/panels/GeneralPage.cpp
@@ -84,6 +84,7 @@ GeneralPage::GeneralPage(Context& ctx, wxWindow* parent)
     m_showWhiteSpaces = editor.get_or("whiteSpace", false);
     m_showLineEndings = editor.get_or("displayEOL", false);
     m_braceHighlight = editor.get_or("braceHighlight", true);
+    m_highlightOccurrences = editor.get_or("highlightOccurrences", true);
     m_syntaxHighlight = editor.get_or("syntaxHighlight", true);
     m_showLineNumbers = editor.get_or("lineNumbers", true);
     m_showRightMargin = editor.get_or("longLine", false);
@@ -110,6 +111,7 @@ void GeneralPage::create() {
             checkBox(m_showWhiteSpaces, tr("dialogs.settings.general.whitespace"));
             checkBox(m_showLineEndings, tr("dialogs.settings.general.lineEndings"));
             checkBox(m_braceHighlight, tr("dialogs.settings.general.braceHighlight"));
+            checkBox(m_highlightOccurrences, tr("dialogs.settings.general.highlightOccurrences"));
             spinCtrl(m_edgeColumn, tr("dialogs.settings.general.rightMarginWidth"), 1, 200, {});
             hbox({ .alignment = SmartBoxSizer::Alignment::Center, .margin = false }, [&] {
                 text(tr("dialogs.settings.general.encoding"), { .expand = false });
@@ -183,6 +185,7 @@ void GeneralPage::apply() {
     editor["whiteSpace"] = m_showWhiteSpaces;
     editor["displayEOL"] = m_showLineEndings;
     editor["braceHighlight"] = m_braceHighlight;
+    editor["highlightOccurrences"] = m_highlightOccurrences;
     editor["syntaxHighlight"] = m_syntaxHighlight;
     editor["lineNumbers"] = m_showLineNumbers;
     editor["longLine"] = m_showRightMargin;